Site Locations: Hands-on Lab: Interactive Visual Analytics with Folium lab.
Hands-on Lab: Interactive Visual Analytics with Folium lab. The launch success rate may depend on many factors such as payload mass, orbit type, and so on. It may also depend on the location and proximities of a launch site, i.e., the initial position of rocket trajectories. Finding an optimal location for building a launch site certainly involves many factors and hopefully we could discover some of the factors by analyzing the existing launch site locations.